+++ Sources/include/HTMLDocument+Private.h
@@ -1,0 +1,11 @@
+//  HTMLDocument+Private.h
+//
+//  Public domain. https://github.com/nolanw/HTMLReader
+
+#import "HTMLDocument.h"
+
+@interface HTMLDocument (Private)
+
+@property (nonatomic) NSStringEncoding parsedStringEncoding;
+
+@end